Move ViewImageConfig extensions into its own files
