+++ Sources/SnapshotKit/Common/ViewImageConfig+tvOS.swift
@@ -1,0 +1,14 @@
+#if os(tvOS)
+public extension ViewImageConfig {
+    static let tv = ViewImageConfig(
+        safeArea: .init(top: 60, left: 90, bottom: 60, right: 90),
+        size: .init(width: 1920, height: 1080),
+        traits: .init()
+    )
+    static let tv4K = ViewImageConfig(
+        safeArea: .init(top: 120, left: 180, bottom: 120, right: 180),
+        size: .init(width: 3840, height: 2160),
+        traits: .init()
+    )
+}
+#endif
